# api/docs/swagger.yaml
definitions:
  models.Basket:
    properties:
      items:
        items:
          $ref: '#/definitions/models.BasketItem'
        type: array
    type: object
  models.BasketItem:
    properties:
      id:
        type: integer
      itemId:
        type: integer
      note:
        type: string
    type: object
...
  models.Dish:
    type: object
...
info:
  contact: {}
paths:
  /api/v1/baskets/items/:
    post:
      description: Gets data from the request body and item information from the database,
        calculates the total cost, and returns the basket items
      parameters:
      - description: JSON object representing the basket
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Basket'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Adds items to basket
      tags:
      - basket
  /api/v1/baskets/pay/{table}:
    post:
      description: Gets item information from basket and table number, calculates
        total cost, adds orders & bill to database
      parameters:
      - description: JSON object representing the basket
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Basket'
      - description: table number
        in: path
        name: table
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: Posts table's bill based on basket
      tags:
      - basket
...
  /api/v1/dishes:
    put:
      description: Puts an dish, if it does not exist it is added otherwise it updates
        the order of the same ID
      parameters:
      - description: JSON object of an order to be added or updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Dish'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Put a dish
      tags:
      - dishes
...
  /api/v1/menu/dishes:
    patch:
      description: Updates the dish on the ID
      parameters:
      - description: JSON object of an order to be added or updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Dish'
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Updates a dish
      tags:
      - dishes
  /api/v1/menu/dishes/:
    get:
      description: Returns the json of all dishes
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Lists all the dishes
      tags:
      - dishes
    post:
      description: Posts a dish into the database
      parameters:
      - description: JSON object of an order to be added or updated
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Dish'
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Posts a dish
      tags:
      - dishes
  /api/v1/menu/dishes/{id}:
    delete:
      description: Deletes the dish on the ID
      parameters:
      - description: ID of the dish
        in: path
        name: id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Deletes a dish
      tags:
      - dishes
    get:
      description: Returns the dish based on its ID
      parameters:
      - description: ID of the dish
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Gets a dish
      tags:
      - dishes
...
swagger: "2.0"